+++ CHANGELOG.md
@@ -11,6 +11,14 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 ### Changed
 
 - Split up kill, suicide and teamkill in the round end screen to make it more clear
+- Added a new in-game player model selector
+  - Added new convars that can change the way playermodels are selected (these can be found in the gamemode menu)
+  - Added a new ConVar `ttt2_use_custom_models` (def: 0) to enable the custom player model selector
+
+### Breaking Changes
+
+- Removed the (unused?) ConVar `ttt2_custom_models`
+- Removed the function `GetRandomPlayerModel()`, use `playermodels.GetRandomPlayerModel()` instead
 
 ## [v0.9.3b](https://github.com/TTT-2/TTT2/tree/v0.9.3b) (2021-09-25)